+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/audio/DecoderAudioRenderer.java
@@ -165,6 +165,7 @@ public abstract class DecoderAudioRenderer<
   private long outputStreamOffsetUs;
   private final long[] pendingOutputStreamOffsetsUs;
   private int pendingOutputStreamOffsetCount;
+  private boolean hasPendingReportedSkippedSilence;
 
   public DecoderAudioRenderer() {
     this(/* eventHandler= */ null, /* eventListener= */ null);
@@ -578,6 +579,13 @@ public abstract class DecoderAudioRenderer<
   }
 
   @Override
+  public boolean hasSkippedSilenceSinceLastCall() {
+    boolean hasPendingReportedSkippedSilence = this.hasPendingReportedSkippedSilence;
+    this.hasPendingReportedSkippedSilence = false;
+    return hasPendingReportedSkippedSilence;
+  }
+
+  @Override
   public void setPlaybackParameters(PlaybackParameters playbackParameters) {
     audioSink.setPlaybackParameters(playbackParameters);
   }
@@ -606,6 +614,7 @@ public abstract class DecoderAudioRenderer<
     audioSink.flush();
 
     currentPositionUs = positionUs;
+    hasPendingReportedSkippedSilence = false;
     allowPositionDiscontinuity = true;
     inputStreamEnded = false;
     outputStreamEnded = false;
@@ -630,6 +639,7 @@ public abstract class DecoderAudioRenderer<
     inputFormat = null;
     audioTrackNeedsConfigure = true;
     setOutputStreamOffsetUs(C.TIME_UNSET);
+    hasPendingReportedSkippedSilence = false;
     try {
       setSourceDrmSession(null);
       releaseDecoder();
@@ -827,6 +837,11 @@ public abstract class DecoderAudioRenderer<
     @Override
     public void onPositionDiscontinuity() {
       DecoderAudioRenderer.this.onPositionDiscontinuity();
+    }
+
+    @Override
+    public void onSilenceSkipped() {
+      hasPendingReportedSkippedSilence = true;
     }
 
     @Override