+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/audio/DefaultAudioSink.java
@@ -96,6 +96,16 @@ public final class DefaultAudioSink implements AudioSink {
    */
   private static final int AUDIO_TRACK_SMALLER_BUFFER_RETRY_SIZE = 1_000_000;
 
+  /** The minimum duration of the skipped silence to be reported as discontinuity. */
+  private static final int MINIMUM_REPORT_SKIPPED_SILENCE_DURATION_US = 1_000_000;
+
+  /**
+   * The delay of reporting the skipped silence, during which the default audio sink checks if there
+   * is any further skipped silence that is close to the delayed silence. If any, the further
+   * skipped silence will be concatenated to the delayed one.
+   */
+  private static final int REPORT_SKIPPED_SILENCE_DELAY_MS = 100;
+
   /**
    * Thrown when the audio track has provided a spurious timestamp, if {@link
    * #failOnSpuriousAudioTimestamp} is set.
@@ -542,6 +552,9 @@ public final class DefaultAudioSink implements AudioSink {
   private boolean offloadDisabledUntilNextConfiguration;
   private boolean isWaitingForOffloadEndOfStreamHandled;
   @Nullable private Looper playbackLooper;
+  private long skippedOutputFrameCountAtLastPosition;
+  private long accumulatedSkippedSilenceDurationUs;
+  private @MonotonicNonNull Handler reportSkippedSilenceHandler;
 
   @RequiresNonNull("#1.audioProcessorChain")
   private DefaultAudioSink(Builder builder) {
@@ -1443,6 +1456,11 @@ public final class DefaultAudioSink implements AudioSink {
     }
     writeExceptionPendingExceptionHolder.clear();
     initializationExceptionPendingExceptionHolder.clear();
+    skippedOutputFrameCountAtLastPosition = 0;
+    accumulatedSkippedSilenceDurationUs = 0;
+    if (reportSkippedSilenceHandler != null) {
+      checkNotNull(reportSkippedSilenceHandler).removeCallbacksAndMessages(null);
+    }
   }
 
   @Override
@@ -1645,8 +1663,28 @@ public final class DefaultAudioSink implements AudioSink {
   }
 
   private long applySkipping(long positionUs) {
-    return positionUs
-        + configuration.framesToDurationUs(audioProcessorChain.getSkippedOutputFrameCount());
+    long skippedOutputFrameCountAtCurrentPosition =
+        audioProcessorChain.getSkippedOutputFrameCount();
+    long adjustedPositionUs =
+        positionUs + configuration.framesToDurationUs(skippedOutputFrameCountAtCurrentPosition);
+    if (skippedOutputFrameCountAtCurrentPosition > skippedOutputFrameCountAtLastPosition) {
+      long silenceDurationUs =
+          configuration.framesToDurationUs(
+              skippedOutputFrameCountAtCurrentPosition - skippedOutputFrameCountAtLastPosition);
+      skippedOutputFrameCountAtLastPosition = skippedOutputFrameCountAtCurrentPosition;
+      handleSkippedSilence(silenceDurationUs);
+    }
+    return adjustedPositionUs;
+  }
+
+  private void handleSkippedSilence(long silenceDurationUs) {
+    accumulatedSkippedSilenceDurationUs += silenceDurationUs;
+    if (reportSkippedSilenceHandler == null) {
+      reportSkippedSilenceHandler = new Handler(Looper.myLooper());
+    }
+    reportSkippedSilenceHandler.removeCallbacksAndMessages(null);
+    reportSkippedSilenceHandler.postDelayed(
+        this::maybeReportSkippedSilence, /* delayMillis= */ REPORT_SKIPPED_SILENCE_DELAY_MS);
   }
 
   private boolean isAudioTrackInitialized() {
@@ -2223,6 +2261,16 @@ public final class DefaultAudioSink implements AudioSink {
     public void clear() {
       pendingException = null;
     }
+  }
+
+  private void maybeReportSkippedSilence() {
+    if (accumulatedSkippedSilenceDurationUs >= MINIMUM_REPORT_SKIPPED_SILENCE_DURATION_US) {
+      // If the existing silence is already long enough, report the silence
+      listener.onSilenceSkipped();
+    }
+    // Reset the accumulated silence anyway as the later silences are far from the current one
+    // and should be treated separately.
+    accumulatedSkippedSilenceDurationUs = 0;
   }
 
   @RequiresApi(23)