+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/util/EventLogger.java
@@ -707,6 +707,8 @@ public class EventLogger implements AnalyticsListener {
         return "SKIP";
       case Player.DISCONTINUITY_REASON_INTERNAL:
         return "INTERNAL";
+      case Player.DISCONTINUITY_REASON_SILENCE_SKIP:
+        return "SILENCE_SKIP";
       default:
         return "?";
     }